requirements.txt
scrapy==0.14.4
python-dateutil==1.5
